requirements.txt
aiogram>=3.7.0
Telethon>=1.36.0
cryptg>=0.5.0
yt-dlp>=2025.1.0
python-dotenv>=1.0.1
aiohttp>=3.10.0
